(BSR)[PRO] refactor: simplify adapter call
