+++ pro/src/pages/Bookings/adapters/index.ts
@@ -1,4 +1,2 @@
 export { getFilteredBookingsRecapAdapter } from './getFilteredBookingsRecapAdapter'
-export { getBookingsCSVFileAdapter } from './getBookingsCSVFileAdapter'
-export { getBookingsXLSFileAdapter } from './getBookingsXLSFileAdapter'
 export { getUserHasBookingsAdapter } from './getUserHasBookingsAdapter'
