+++ pro/src/screens/Bookings/Bookings.tsx
@@ -9,8 +9,6 @@ import {
 import NoData from 'components/NoData'
 import { DEFAULT_PRE_FILTERS } from 'core/Bookings/constants'
 import {
-  GetBookingsCSVFileAdapter,
-  GetBookingsXLSFileAdapter,
   GetFilteredBookingsRecapAdapter,
   GetFilteredCollectiveBookingsRecapAdapter,
   GetUserHasBookingsAdapter,
@@ -35,13 +33,11 @@ import Titles from 'ui-kit/Titles/Titles'
 import { stringify } from '../../utils/query-string'
 
 import BookingsRecapTable from './BookingsRecapTable'
-import PreFilters from './PreFilters'
+import { PreFilters } from './PreFilters/PreFilters'
 
 interface BookingsProps {
   locationState?: { statuses: string[] }
   audience: Audience
-  getBookingsCSVFileAdapter: GetBookingsCSVFileAdapter
-  getBookingsXLSFileAdapter: GetBookingsXLSFileAdapter
   getFilteredBookingsRecapAdapter:
     | GetFilteredBookingsRecapAdapter
     | GetFilteredCollectiveBookingsRecapAdapter
@@ -56,8 +52,6 @@ const Bookings = <
 >({
   locationState,
   audience,
-  getBookingsCSVFileAdapter,
-  getBookingsXLSFileAdapter,
   getFilteredBookingsRecapAdapter,
   getUserHasBookingsAdapter,
   getVenuesAdapter,
@@ -263,8 +257,7 @@ const Bookings = <
       <PreFilters
         appliedPreFilters={appliedPreFilters}
         applyPreFilters={applyPreFilters}
-        getBookingsCSVFileAdapter={getBookingsCSVFileAdapter}
-        getBookingsXLSFileAdapter={getBookingsXLSFileAdapter}
+        audience={audience}
         hasResult={bookings.length > 0}
         isFiltersDisabled={!hasBooking}
         isLocalLoading={isLocalLoading}
