--- pro/src/screens/Bookings/PreFilters/index.ts
@@ -1,1 +1,0 @@
-export { default } from './PreFilters'
